requirements.txt
Flask>=2.0
pyyaml